DASHBOARD/requirements.txt
asgiref==3.4.1
Django==4.0.1
django-filter==21.1
numpy==1.22.2
pandas==1.4.0
python-dateutil==2.8.2
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
